enroll-stripe: Add feature to request payment upon enrollment
The feature is activated by setting `REQUIRE_PAYMENT_TO_ENROLL_PRICE` to
True in `MITX_FEATURES`.

When activated, when a student attempts to enroll on a given course
session, a Stripe payment form is presented to him, and he only gets
to actually enroll once the payment is cleared. Staff is unaffected.

Configuration parameters:

* `REQUIRE_PAYMENT_TO_ENROLL_PRICE`: price in cents per course session
  (default = 12500)
* `REQUIRE_PAYMENT_TO_ENROLL_CURRENCY`: currency for price (default =
  'usd')
* `STRIPE_KEY_PUBLIC` and `STRIPE_KEY_SECRET`: Stripe API keys