registration-optional: Allow to hide some optional registration fields
Allow to control which of the optional fields from the registration page
are being displayed to the user. Added the configuration variable
`REGISTRATION_OPTIONAL_FIELDS` to control which ones. The default value
shows all the fields as currently; to not display some of the fields,
set the configuration variable with only the fields you want to display:

```
REGISTRATION_OPTIONAL_FIELDS = [
    'level_of_education',
    'gender',
    'year_of_birth',
    'mailing_address',
    'goals'
]
```